requirements.txt
streamlit==1.5.0
matplotlib==3.5.1
mysql-connector-python==8.0.27
seaborn==0.11.2
pandas==1.4.0